chore: move utils and types to separate files
