+++ web/src/components/pages/guidelines/types.ts
@@ -1,0 +1,5 @@
+export type SidebarContentProps = {
+  id: string;
+  contentVisible: boolean;
+  toggleVisibleSection: (id: string) => void;
+};
