+++ web/src/components/pages/guidelines/utils.ts
@@ -1,0 +1,31 @@
+import { voiceGuidelinesSections } from './constants';
+import { SidebarContentProps } from './types';
+
+type Section = {
+  id: string;
+  component: React.FC<SidebarContentProps>;
+  visible: boolean;
+};
+
+export type handleToggleVisibleSectionParams = {
+  visibleSections: Section[];
+  setVisibleSections: React.Dispatch<
+    React.SetStateAction<typeof voiceGuidelinesSections>
+  >;
+  id: string;
+};
+
+export const handleToggleVisibleSection = ({
+  visibleSections,
+  setVisibleSections,
+  id,
+}: handleToggleVisibleSectionParams) => {
+  const newSections = visibleSections.map(section => {
+    if (section.id === id) {
+      return { ...section, visible: !section.visible };
+    }
+    return section;
+  });
+
+  setVisibleSections(newSections);
+};
